+++ packages/cms/lib/modules/choices-guide-widgets/lib/create-config.js
@@ -30,9 +30,14 @@ module.exports = function createConfig(widget, data, jwt, apiUrl, loginUrl) {
     },
     beforeUrl: widget.beforeUrl && data.siteUrl + widget.beforeUrl,
     afterUrl: widget.afterUrl && data.siteUrl + widget.afterUrl,
-    imageserver: {
-			process: '/image',
-			fetch: '/image',
+    image: {
+      server: {
+				process: '/image',
+				fetch: '/image',
+        srcExtension: '/:/rs=w:[[width]],h:[[height]];cp=w:[[width]],h:[[height]]',
+      },
+      aspectRatio: widget.imageAspectRatio || '16x9',
+      allowMultipleImages: false,
     },
   }